+++ esphome/core/hal.h
@@ -30,15 +30,11 @@
 
 namespace esphome {
 
-// ESP8266 inlines delayMicroseconds() and arch_feed_wdt() in hal/hal_esp8266.h;
-// every other platform keeps them out-of-line in components/<platform>/core.cpp.
-#ifndef USE_ESP8266
-void delayMicroseconds(uint32_t us);  // NOLINT(readability-identifier-naming)
-void arch_feed_wdt();
-#endif
+// Cross-platform declarations. delayMicroseconds(), arch_feed_wdt(),
+// arch_get_cpu_cycle_count() vary per platform (some inline, some
+// out-of-line) so they live in hal/hal_<platform>.h.
 void __attribute__((noreturn)) arch_restart();
 void arch_init();
-uint32_t arch_get_cpu_cycle_count();
 uint32_t arch_get_cpu_freq_hz();
 
 #ifndef USE_ESP8266
